pyproject.toml
[tool.poetry]
name = "personal_portfolio"
version = "0.1.0"
description = "My personal portfolio"
authors = ["Keith Riordan <keith.riordan@gmail.com>"]

[tool.poetry.dependencies]
python = "3.12.1"
Flask = "^3.0"
Flask-SQLAlchemy = "^3.0"
boto3 = "^1.24.41"
python-dotenv = "^0.21.0"
requests = "^2.28.1"
PyYAML = "^6.0"
Flask-WTF = "^1.0.1"
email-validator = "^1.2.1"
psycopg2 = "^2.9.3"
flask-login = "^0.6.3"
flask-migrate = "^4.0.5"
sendgrid = "^6.11.0"
flask-talisman = "^1.1.0"
flask-apscheduler = "^1.13.1"
google-api-python-client = "^2.138.0"
google-auth = "^2.32.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"

[tool.poetry.dev-dependencies]
djlint = "^1.34.0"
faker = "^20.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
